BIG update!
I added tons of new stuff to make up for how long I was gone! Of course, most of this was behind the scenes, but still! I did stuff!

First, I added a new article about the execrable Elder Scrolls VI "trailer" that was shown at E3.

Second, I deleted all the fonts that weren't in the fonts folder (this is something that was supposed to be done in the previous update, but I forgot).

Third, I said, "Wait, what if the readers click on my links, watch the YouTube videos, and then never come back to the website?" I didn't want this to happen, so I added code to <head> to make every link on the website except the nav bar open in a new tab. That way, they'll never forget where they were. I know, I'm a genius.

Finally, I added an article in News informing the readers that all links would be opening in a new tab from now on. Because they deserve to know.